size

+++ 0x00-hello_world/6-size.c
@@ -1,4 +1,5 @@
 #include <stdio.h>
+
 /**
  * main - print out sizes of data types in C
  * code by Nneoma
@@ -6,16 +7,16 @@
 */
 int main(void)
 {
-	char a;
-	int b;
-	long int c;
-	long long int d;
-	float f;
+char a;
+int b;
+long int c;
+long long int d;
+float f;
 
-	printf("Size of a char: %lu bytes(s)\n", (unsigned long)sizeof(a));
-	printf("Size of an int: %lu bytes(s)\n", (unsigned long)sizeof(b));
-	printf("Size of a long int: %lu bytes(s)\n", (unsigned long)sizeof(c));
-	printf("Size of a long long int: %lu bytes(s)\n", (unsigned long)sizeof(d));
-	printf("Size of a float: %lu bytes(s)\n", (unsigned long)sizeof(f));
+	printf("Size of a char: %lu bytes(s)\n", sizeof(a));
+	printf("Size of an int: %lu bytes(s)\n", sizeof(b));
+	printf("Size of a long int: %lu bytes(s)\n", sizeof(c));
+	printf("Size of a long long int: %lu bytes(s)\n", sizeof(d));
+	printf("Size of a float: %lu bytes(s)\n", sizeof(f));
 	return (0);
 }
